+++ DirectN.Extensions/Utilities/AllocPstr.cs
@@ -5,16 +5,16 @@ public class AllocPstr : Pwstr
     public AllocPstr(int sizeInBytes)
     {
         ArgumentOutOfRangeException.ThrowIfNegative(sizeInBytes);
-        SizeInBytes = sizeInBytes;
+        SizeInBytes = (uint)sizeInBytes;
         Value = sizeInBytes > 0 ? Marshal.AllocCoTaskMem(sizeInBytes) : 0;
     }
 
     public AllocPstr(uint sizeInBytes)
     {
-        SizeInBytes = (int)sizeInBytes;
+        SizeInBytes = sizeInBytes;
         Value = sizeInBytes > 0 ? Marshal.AllocCoTaskMem((int)sizeInBytes) : 0;
     }
 
-    public int SizeInBytes { get; }
-    public int SizeInChars => SizeInChars;
+    public uint SizeInBytes { get; }
+    public uint SizeInChars => SizeInChars;
 }
