+++ DirectN.Extensions/Utilities/AllocPwstr.cs
@@ -5,16 +5,16 @@ public class AllocPwstr : Pwstr
     public AllocPwstr(int sizeInBytes)
     {
         ArgumentOutOfRangeException.ThrowIfNegative(sizeInBytes);
-        SizeInBytes = sizeInBytes;
+        SizeInBytes = (uint)sizeInBytes;
         Value = sizeInBytes > 0 ? Marshal.AllocCoTaskMem(sizeInBytes) : 0;
     }
 
     public AllocPwstr(uint sizeInBytes)
     {
-        SizeInBytes = (int)sizeInBytes;
+        SizeInBytes = sizeInBytes;
         Value = sizeInBytes > 0 ? Marshal.AllocCoTaskMem((int)sizeInBytes) : 0;
     }
 
-    public int SizeInBytes { get; }
-    public int SizeInChars => SizeInChars / 2;
+    public uint SizeInBytes { get; }
+    public uint SizeInChars => SizeInChars / 2;
 }
